First round of notes for Kevin
