+++ .envrc
@@ -1,0 +1,2 @@
+
+source .venv/bin/activate
